[ACTIVEMODEL] [¡BREAKING!] Wrap search results in Item class, not the actual model class [Closes #11]
Due to the problem with indexing and retrieving ActiveRecord models with associations (such as Article has_many :comments),
do not wrap the results in the model class, but in the Tire::Configuration.wrapper class.

See the https://github.com/karmi/tire/issues/10 issue for explanation.

This commit breaks the old behaviour and usage in Rails, since Results::Item is not (yet) ActiveModel compatible.